(refactor) No need for separate shared folder, and remove dead code
This has also revealed that there is unnecessary AWS authentication code. This
is a remnant of when AWS Elasticsearch was used. It is now in PaaS, which uses
basic auth.
