+++ core/app/app_server.py
@@ -3,14 +3,6 @@ import time
 
 from aiohttp import web
 
-from ..shared.logger import (
-    logged,
-    get_child_logger,
-)
-from ..shared.utils import (
-    random_url_safe,
-)
-
 from .app_elasticsearch import (
     es_search,
     es_search_existing_scroll,
@@ -20,8 +12,13 @@ from .app_elasticsearch import (
 from .app_hawk import (
     authenticate_hawk_header,
 )
+from .app_logger import (
+    logged,
+    get_child_logger,
+)
 from .app_utils import (
     get_child_context,
+    random_url_safe,
 )
 from .app_redis import (
     set_private_scroll_id,
@@ -249,3 +246,68 @@ def json_response(data, status):
     return web.json_response(data, status=status, headers={
         'Server': 'activity-stream'
     })
+
+
+def server_logger(logger):
+
+    @web.middleware
+    async def _server_logger(request, handler):
+        child_logger = get_child_logger(logger, random_url_safe(8))
+        request['logger'] = child_logger
+        child_logger.debug('Receiving request (%s) (%s %s HTTP/%s.%s) (%s) (%s)', *(
+            (
+                request.remote,
+                request.method,
+                request.path_qs,
+            ) +
+            request.version +
+            (
+                request.headers.get('User-Agent', '-'),
+                request.headers.get('X-Forwarded-For', '-'),
+            )
+        ))
+
+        with logged(child_logger, 'Processing request', []):
+            response = await handler(request)
+
+        child_logger.debug(
+            'Sending Response (%s) (%s)',
+            response.status, response.content_length,
+        )
+
+        return response
+
+    return _server_logger
+
+
+def authenticate_by_ip(incorrect, ip_whitelist):
+
+    @web.middleware
+    async def _authenticate_by_ip(request, handler):
+        if 'X-Forwarded-For' not in request.headers:
+            request['logger'].warning(
+                'Failed authentication: no X-Forwarded-For header passed'
+            )
+            raise web.HTTPUnauthorized(text=incorrect)
+
+        # PaaS appends 2 IPs, where the IP connected from is the first of the two
+        ip_addesses = request.headers['X-Forwarded-For'].split(',')
+        if len(ip_addesses) < 2:
+            request['logger'].warning(
+                'Failed authentication: the X-Forwarded-For header does not '
+                'contain enough IP addresses'
+            )
+            raise web.HTTPUnauthorized(text=incorrect)
+
+        remote_address = ip_addesses[-2].strip()
+
+        if remote_address not in ip_whitelist:
+            request['logger'].warning(
+                'Failed authentication: the IP address derived from the '
+                'X-Forwarded-For header is not in the whitelist'
+            )
+            raise web.HTTPUnauthorized(text=incorrect)
+
+        return await handler(request)
+
+    return _authenticate_by_ip
