+++ core/app/app_utils.py
@@ -1,12 +1,16 @@
 import asyncio
 import collections
+import itertools
+import json
 import logging
+import secrets
 import signal
+import string
 import sys
 
 import aiohttp
 
-from ..shared.logger import (
+from .app_logger import (
     logged,
     get_child_logger,
 )
@@ -100,6 +104,125 @@ def http_429_retry_after(coroutine):
                 await asyncio.sleep(int(client_error.headers['Retry-After']))
 
     return _http_429_retry_after
+
+
+def random_url_safe(count):
+    return ''.join(secrets.choice(string.ascii_lowercase + string.digits) for _ in range(count))
+
+
+def get_common_config(env):
+    vcap_services = json.loads(env['VCAP_SERVICES'])
+    es_uri = vcap_services['elasticsearch'][0]['credentials']['uri']
+    redis_uri = vcap_services['redis'][0]['credentials']['uri']
+    sentry = {
+        'dsn': env['SENTRY_DSN'],
+        'environment': env['SENTRY_ENVIRONMENT'],
+    }
+    return es_uri, redis_uri, sentry
+
+
+def normalise_environment(key_values):
+    ''' Converts denormalised dict of (string -> string) pairs, where the first string
+        is treated as a path into a nested list/dictionary structure
+
+        {
+            "FOO__1__BAR": "setting-1",
+            "FOO__1__BAZ": "setting-2",
+            "FOO__2__FOO": "setting-3",
+            "FOO__2__BAR": "setting-4",
+            "FIZZ": "setting-5",
+        }
+
+        to the nested structure that this represents
+
+        {
+            "FOO": [{
+                "BAR": "setting-1",
+                "BAZ": "setting-2",
+            }, {
+                "BAR": "setting-3",
+                "BAZ": "setting-4",
+            }],
+            "FIZZ": "setting-5",
+        }
+
+        If all the keys for that level parse as integers, then it's treated as a list
+        with the actual keys only used for sorting
+
+        This function is recursive, but it would be extremely difficult to hit a stack
+        limit, and this function would typically by called once at the start of a
+        program, so efficiency isn't too much of a concern.
+    '''
+
+    # Separator is chosen to
+    # - show the structure of variables fairly easily;
+    # - avoid problems, since underscores are usual in environment variables
+    separator = '__'
+
+    def get_first_component(key):
+        return key.split(separator)[0]
+
+    def get_later_components(key):
+        return separator.join(key.split(separator)[1:])
+
+    without_more_components = {
+        key: value
+        for key, value in key_values.items()
+        if not get_later_components(key)
+    }
+
+    with_more_components = {
+        key: value
+        for key, value in key_values.items()
+        if get_later_components(key)
+    }
+
+    def grouped_by_first_component(items):
+        def by_first_component(item):
+            return get_first_component(item[0])
+
+        # groupby requires the items to be sorted by the grouping key
+        return itertools.groupby(
+            sorted(items, key=by_first_component),
+            by_first_component,
+        )
+
+    def items_with_first_component(items, first_component):
+        return {
+            get_later_components(key): value
+            for key, value in items
+            if get_first_component(key) == first_component
+        }
+
+    nested_structured_dict = {
+        **without_more_components, **{
+            first_component: normalise_environment(
+                items_with_first_component(items, first_component))
+            for first_component, items in grouped_by_first_component(with_more_components.items())
+        }}
+
+    def all_keys_are_ints():
+        def is_int(string_to_test):
+            try:
+                int(string_to_test)
+                return True
+            except ValueError:
+                return False
+
+        return all([is_int(key) for key, value in nested_structured_dict.items()])
+
+    def list_sorted_by_int_key():
+        return [
+            value
+            for key, value in sorted(
+                nested_structured_dict.items(),
+                key=lambda key_value: int(key_value[0])
+            )
+        ]
+
+    return \
+        list_sorted_by_int_key() if all_keys_are_ints() else \
+        nested_structured_dict
 
 
 def main(run_application_coroutine):
